Renamed new_test.py to test.py and deleted the old test.py
